monitoring/pyproject.toml
[tool.poetry]
name = "uss"
version = "0.1.0"
description = ""
authors = ["Sebastian RODRIGUEZ <sebastian.rodriguez@rmit.edu.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
behave = "^1.2.6"
pytest = "^8.3.3"
pytest-watch = "^4.2.0"
python-logging-loki = "^0.3.1"
coloredlogs = "^15.0.1"
allure-behave = "^2.13.5"
prometheus-client = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.basedpyright]
typeCheckingMode = "standard"
